% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_genotypes.R
\name{add_genotypes}
\alias{add_genotypes}
\title{Use Genomic Information}
\usage{
add_genotypes(path, G_geno)
}
\arguments{
\item{path}{String}

\item{G_geno}{A number or diag matrix}
}
\description{
Use Genomic Information
}
\examples{
G3 = diag(3)
geno_path = "D:/JWASr/data/genotypes.txt"
add_genotypes(geno_path, G3)


}
