% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\title{Build model}
\usage{
build_model(model_equation, R)
}
\arguments{
\item{model_equation}{the model equation}

\item{R}{the residual value}
}
\description{
Build model
}
\examples{
model_equation = "y1 = intercept + x1"
R = 1.0
model = build_model(model_equation, R)

}
