% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwasr_setup.R
\name{jwasr_setup}
\alias{jwasr_setup}
\title{Set up}
\usage{
jwasr_setup(path_libjulia)
}
\arguments{
\item{path_libjulia}{path of libjulia}
}
\description{
This function initializes Julia and the JWAS.jl package
The first time will be long since it includes precompilation.
}
\examples{
path_libjulia = "C:/Users/ztjsw/AppData/Local/Julia-0.7.0/bin/libjulia.dll"
jwasr_setup(path_libjulia)

}
