% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwasr_setup_win.R
\name{jwasr_setup_win}
\alias{jwasr_setup_win}
\title{Set up for windows user}
\usage{
jwasr_setup_win(path_libjulia)
}
\arguments{
\item{path_libjulia}{path of libjulia}
}
\description{
This function initializes Julia and the JWAS.jl package
}
\examples{
path_libjulia = "C:/Users/ztjsw/AppData/Local/Julia-0.7.0/bin/libjulia.dll"
jwasr_setup_win(path_libjulia)

}
