% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_random.R
\name{set_random}
\alias{set_random}
\title{Set Random or Fixed Effects}
\usage{
set_random(xi, G, pedigree = FALSE)
}
\arguments{
\item{xi}{String}

\item{G}{number or diag matrix}

\item{pedigree}{FALSE or TRUE}
}
\description{
Set Random or Fixed Effects
}
