% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwasr_setup_mac.R
\name{jwasr_setup}
\alias{jwasr_setup}
\title{Set up}
\usage{
jwasr_setup()
}
\description{
This function initializes Julia and the JWAS.jl package
The first time will be long since it includes precompilation.
}
\details{
jwasr_setup()
}
