% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCMC.R
\name{runMCMC}
\alias{runMCMC}
\title{Run the MCMC model}
\usage{
runMCMC(model, mydata, chain_length = 1000,
  output_samples_frequency = 0)
}
\arguments{
\item{model}{the model built before}

\item{mydata}{the data upload before}
}
\description{
Run the MCMC model
}
\examples{
model_equation = "y1 = intercept + x1"
R = 1.0
model = build_model(model_equation, R)
runMCMC(model, phenotypes)

}
